+++ auth-service/src/main/java/org/example/auth/config/security/role/RoleChecker.java
@@ -1,7 +1,7 @@
 package org.example.auth.config.security.role;
 
 public interface RoleChecker {
-    boolean isRoleAvailable(Integer personId);
+    boolean isRoleAvailable(Long personId);
 
     String getType();
 }
