+++ course-registration-service/src/main/java/org/example/courseregistration/dto/courserequest/CreateKeeperRejectionDto.java
@@ -11,5 +11,5 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 public class CreateKeeperRejectionDto {
     @NotNull
-    private Integer reasonId;
+    private Long reasonId;
 }
