+++ course-registration-service/src/main/java/org/example/courseregistration/dto/progress/CoursesStateDto.java
@@ -7,14 +7,14 @@ import java.util.Set;
 
 @Data
 public class CoursesStateDto {
-    private Integer personId;
+    private Long personId;
     private String firstName;
     private String lastName;
     private String patronymic;
     @JsonProperty("openedSystems")
-    private Set<Integer> openedCourses;
+    private Set<Long> openedCourses;
     @JsonProperty("studiedSystems")
     private Set<CourseWithProgressDto> studiedCourses;
     @JsonProperty("closedSystems")
-    private Set<Integer> closedCourses;
+    private Set<Long> closedCourses;
 }
