+++ course-registration-service/src/main/java/org/example/courseregistration/exception/classes/request/RequestAlreadyClosedException.java
@@ -1,7 +1,7 @@
 package org.example.courseregistration.exception.classes.request;
 
 public class RequestAlreadyClosedException extends RuntimeException {
-    public RequestAlreadyClosedException(Integer requestId) {
+    public RequestAlreadyClosedException(Long requestId) {
         super("Запрос " + requestId + " уже был закрыт");
     }
 }
